mcp-server/requirements.txt
fastapi==0.109.2
uvicorn==0.23.2
pydantic==1.10.14
pydantic[email]
azure-cosmos==4.6.0
httpx==0.26.0
python-dotenv==1.0.1
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.21.1
